Move SQL identity certification into core registration
