+++ extension/core_functions/function_list.cpp
@@ -58,11 +58,9 @@ static const StaticFunctionDefinition core_functions[] = {
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(ListCosineDistanceFunAlias),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ListHasAllFunAlias2),
 	DUCKDB_SCALAR_FUNCTION_SET(RightShiftFun),
-	DUCKDB_SCALAR_FUNCTION_SET(AbsOperatorFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ListHasAllFunAlias),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(PowOperatorFunAlias),
 	DUCKDB_SCALAR_FUNCTION(StartsWithOperatorFun),
-	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(AbsFun),
 	DUCKDB_SCALAR_FUNCTION(AcosFun),
 	DUCKDB_SCALAR_FUNCTION(AcoshFun),
 	DUCKDB_SCALAR_FUNCTION_SET(AgeFun),
@@ -360,8 +358,6 @@ static const StaticFunctionDefinition core_functions[] = {
 	DUCKDB_SCALAR_FUNCTION(StructKeysFun),
 	DUCKDB_SCALAR_FUNCTION(StructUpdateFun),
 	DUCKDB_SCALAR_FUNCTION(StructValuesFun),
-	DUCKDB_AGGREGATE_FUNCTION_SET(SumFun),
-	DUCKDB_AGGREGATE_FUNCTION_SET(SumNoOverflowFun),
 	DUCKDB_AGGREGATE_FUNCTION_ALIAS(SumkahanFun),
 	DUCKDB_SCALAR_FUNCTION_SET(SwitchFun),
 	DUCKDB_SCALAR_FUNCTION(TanFun),
