+++ extension/core_functions/scalar/date/date_part.cpp
@@ -10,24 +10,10 @@
 #include "duckdb/common/types/time.hpp"
 #include "duckdb/common/types/timestamp.hpp"
 #include "duckdb/function/scalar/nested_functions.hpp"
-#include "function_identity.hpp"
 #include "duckdb/planner/expression/bound_function_expression.hpp"
 #include "duckdb/execution/expression_executor.hpp"
 
 namespace duckdb {
-
-struct DatePartFunctionIdentity {
-	static void PreserveStatistics(ScalarFunction &function) {
-		FunctionIdentityPreservation::PreserveStatistics(function);
-	}
-	static void PreserveStatistics(ScalarFunctionSet &functions) {
-		functions.ApplyToFunctions([](ScalarFunction &function) {
-			if (function.HasStatisticsCallback()) {
-				PreserveStatistics(function);
-			}
-		});
-	}
-};
 
 namespace {
 DatePartSpecifier GetDateTypePartSpecifier(const string &specifier, const LogicalType &type) {
@@ -2184,7 +2170,6 @@ ScalarFunctionSet GetGenericDatePartFunction(scalar_function_t date_func, scalar
 	operator_set.AddFunction(std::move(timestamp_function));
 	operator_set.AddFunction(ScalarFunction({LogicalType::INTERVAL}, LogicalType::BIGINT, std::move(interval_func)));
 	operator_set.SetFallible();
-	DatePartFunctionIdentity::PreserveStatistics(operator_set);
 	return operator_set;
 }
 
@@ -2204,21 +2189,16 @@ ScalarFunctionSet GetGenericTimePartFunction(const LogicalType &result_type, sca
                                              function_statistics_t time_ns_stats, function_statistics_t timetz_stats) {
 	ScalarFunctionSet operator_set;
 	ScalarFunction date_function({LogicalType::DATE}, result_type, std::move(date_func), nullptr, date_stats);
-	DatePartFunctionIdentity::PreserveStatistics(date_function);
 	operator_set.AddFunction(std::move(date_function));
 	ScalarFunction timestamp_function({LogicalType::TIMESTAMP}, result_type, std::move(ts_func), nullptr, ts_stats);
-	DatePartFunctionIdentity::PreserveStatistics(timestamp_function);
 	operator_set.AddFunction(std::move(timestamp_function));
 	operator_set.AddFunction(ScalarFunction({LogicalType::INTERVAL}, result_type, std::move(interval_func)));
 	ScalarFunction time_function({LogicalType::TIME}, result_type, std::move(time_func), nullptr, time_stats);
-	DatePartFunctionIdentity::PreserveStatistics(time_function);
 	operator_set.AddFunction(std::move(time_function));
 	ScalarFunction time_ns_function({LogicalType::TIME_NS}, result_type, std::move(time_ns_func), nullptr,
 	                                time_ns_stats);
-	DatePartFunctionIdentity::PreserveStatistics(time_ns_function);
 	operator_set.AddFunction(std::move(time_ns_function));
 	ScalarFunction timetz_function({LogicalType::TIME_TZ}, result_type, std::move(timetz_func), nullptr, timetz_stats);
-	DatePartFunctionIdentity::PreserveStatistics(timetz_function);
 	operator_set.AddFunction(std::move(timetz_function));
 	return operator_set;
 }
