+++ extension/core_functions/scalar/math/functions.json
@@ -1,12 +1,6 @@
 [
     {
-        "name": "@",
-        "parameters": "x",
-        "description": "Absolute value",
-        "example": "abs(-17.4)",
-        "type": "scalar_function_set",
-        "struct": "AbsOperatorFun",
-        "aliases": ["abs"]
+        "include": "duckdb/function/scalar/operator_functions.hpp"
     },
     {
         "name": "**",
