+++ extension/core_functions/scalar/math/numeric.cpp
@@ -9,7 +9,6 @@
 #include "duckdb/common/vector_operations/unary_executor.hpp"
 #include "core_functions/scalar/math_functions.hpp"
 #include "duckdb/execution/expression_executor.hpp"
-#include "function_identity.hpp"
 #include "duckdb/main/settings.hpp"
 #include "duckdb/planner/expression/bound_constant_expression.hpp"
 #include "duckdb/planner/expression/bound_function_expression.hpp"
@@ -19,19 +18,6 @@
 #include <type_traits>
 
 namespace duckdb {
-
-struct NumericFunctionIdentity {
-	static void PreserveStatistics(ScalarFunction &function) {
-		FunctionIdentityPreservation::PreserveStatistics(function);
-	}
-	static void PreserveStatistics(ScalarFunctionSet &functions) {
-		functions.ApplyToFunctions([](ScalarFunction &function) {
-			if (function.HasStatisticsCallback()) {
-				PreserveStatistics(function);
-			}
-		});
-	}
-};
 
 template <class ERROR_OP, class IEEE_OP>
 static unique_ptr<FunctionData> BindIEEEFloatingUnary(BindScalarFunctionInput &input) {
@@ -112,177 +98,6 @@ ScalarFunctionSet NextAfterFun::GetFunctions() {
 	next_after_fun.AddFunction(ScalarFunction({LogicalType::FLOAT, LogicalType::FLOAT}, LogicalType::FLOAT,
 	                                          ScalarFunction::BinaryFunction<float, float, float, NextAfterOperator>));
 	return next_after_fun;
-}
-
-//===--------------------------------------------------------------------===//
-// abs
-//===--------------------------------------------------------------------===//
-static unique_ptr<BaseStatistics> PropagateAbsStats(ClientContext &context, FunctionStatisticsInput &input) {
-	auto &child_stats = input.child_stats;
-	auto &expr = input.expr;
-	D_ASSERT(child_stats.size() == 1);
-	// can only propagate stats if the children have stats
-	auto &lstats = child_stats[0];
-	Value new_min, new_max;
-	bool potential_overflow = true;
-	if (NumericStats::HasMinMax(lstats)) {
-		switch (expr.GetReturnType().InternalType()) {
-		case PhysicalType::FLOAT:
-		case PhysicalType::DOUBLE:
-			potential_overflow = false;
-			break;
-		case PhysicalType::INT8:
-			potential_overflow = NumericStats::Min(lstats).GetValue<int8_t>() == NumericLimits<int8_t>::Minimum();
-			break;
-		case PhysicalType::INT16:
-			potential_overflow = NumericStats::Min(lstats).GetValue<int16_t>() == NumericLimits<int16_t>::Minimum();
-			break;
-		case PhysicalType::INT32:
-			potential_overflow = NumericStats::Min(lstats).GetValue<int32_t>() == NumericLimits<int32_t>::Minimum();
-			break;
-		case PhysicalType::INT64:
-			potential_overflow = NumericStats::Min(lstats).GetValue<int64_t>() == NumericLimits<int64_t>::Minimum();
-			break;
-		default:
-			return nullptr;
-		}
-	}
-	if (potential_overflow) {
-		new_min = Value(expr.GetReturnType());
-		new_max = Value(expr.GetReturnType());
-	} else {
-		// no potential overflow
-
-		// compute stats
-		switch (expr.GetReturnType().InternalType()) {
-		case PhysicalType::FLOAT:
-		case PhysicalType::DOUBLE: {
-			auto current_min = NumericStats::Min(lstats).GetValue<double>();
-			auto current_max = NumericStats::Max(lstats).GetValue<double>();
-			if (Value::IsNan(current_min) || Value::IsNan(current_max)) {
-				return nullptr;
-			}
-
-			double min_val, max_val;
-			if (current_min == 0 || current_max == 0) {
-				// Unlike integers, floating point abs cannot be removed for zero: abs(-0.0) clears the sign bit.
-				min_val = AbsOperator::Operation<double, double>(current_min);
-				max_val = AbsOperator::Operation<double, double>(current_max);
-			} else if (current_min < 0 && current_max < 0) {
-				min_val = AbsOperator::Operation<double, double>(current_max);
-				max_val = AbsOperator::Operation<double, double>(current_min);
-			} else if (current_min < 0) {
-				D_ASSERT(current_max >= 0);
-				min_val = 0;
-				max_val = MaxValue(AbsOperator::Operation<double, double>(current_min), current_max);
-			} else {
-				*input.expr_ptr = std::move(input.expr.GetChildrenMutable()[0]);
-				return child_stats[0].ToUnique();
-			}
-			new_min = expr.GetReturnType().InternalType() == PhysicalType::FLOAT
-			              ? Value::FLOAT(static_cast<float>(min_val))
-			              : Value::DOUBLE(min_val);
-			new_max = expr.GetReturnType().InternalType() == PhysicalType::FLOAT
-			              ? Value::FLOAT(static_cast<float>(max_val))
-			              : Value::DOUBLE(max_val);
-			break;
-		}
-		default: {
-			auto current_min = NumericStats::Min(lstats).GetValue<int64_t>();
-			auto current_max = NumericStats::Max(lstats).GetValue<int64_t>();
-
-			int64_t min_val, max_val;
-			if (current_min < 0 && current_max < 0) {
-				// if both min and max are below zero, then min=abs(cur_max) and max=abs(cur_min)
-				min_val = AbsValue(current_max);
-				max_val = AbsValue(current_min);
-			} else if (current_min < 0) {
-				D_ASSERT(current_max >= 0);
-				// if min is below zero and max is above 0, then min=0 and max=max(cur_max, abs(cur_min))
-				min_val = 0;
-				max_val = MaxValue(AbsValue(current_min), current_max);
-			} else {
-				// if both current_min and current_max are > 0, then the abs is a no-op and can be removed entirely
-				*input.expr_ptr = std::move(input.expr.GetChildrenMutable()[0]);
-				return child_stats[0].ToUnique();
-			}
-			new_min = Value::Numeric(expr.GetReturnType(), min_val);
-			new_max = Value::Numeric(expr.GetReturnType(), max_val);
-			break;
-		}
-		}
-		expr.FunctionMutable().SetFunctionCallback(
-		    ScalarFunction::GetScalarUnaryFunction<AbsOperator>(expr.GetReturnType()));
-	}
-	auto stats = NumericStats::CreateEmpty(expr.GetReturnType());
-	NumericStats::SetMin(stats, new_min);
-	NumericStats::SetMax(stats, new_max);
-	stats.CopyValidity(lstats);
-	return stats.ToUnique();
-}
-
-template <class OP>
-static unique_ptr<FunctionData> DecimalUnaryOpBind(BindScalarFunctionInput &input) {
-	auto &bound_function = input.GetBoundFunction();
-	auto &arguments = input.GetArguments();
-	auto decimal_type = arguments[0]->GetReturnType();
-	switch (decimal_type.InternalType()) {
-	case PhysicalType::INT16:
-		bound_function.SetFunctionCallback(ScalarFunction::GetScalarUnaryFunction<OP>(LogicalTypeId::SMALLINT));
-		break;
-	case PhysicalType::INT32:
-		bound_function.SetFunctionCallback(ScalarFunction::GetScalarUnaryFunction<OP>(LogicalTypeId::INTEGER));
-		break;
-	case PhysicalType::INT64:
-		bound_function.SetFunctionCallback(ScalarFunction::GetScalarUnaryFunction<OP>(LogicalTypeId::BIGINT));
-		break;
-	default:
-		bound_function.SetFunctionCallback(ScalarFunction::GetScalarUnaryFunction<OP>(LogicalTypeId::HUGEINT));
-		break;
-	}
-	bound_function.GetArguments()[0] = decimal_type;
-	bound_function.SetReturnType(decimal_type);
-	return nullptr;
-}
-
-ScalarFunctionSet AbsOperatorFun::GetFunctions() {
-	ScalarFunctionSet abs;
-	for (auto &type : LogicalType::Numeric()) {
-		switch (type.id()) {
-		case LogicalTypeId::DECIMAL:
-			abs.AddFunction(ScalarFunction({type}, type, nullptr, DecimalUnaryOpBind<AbsOperator>));
-			break;
-		case LogicalTypeId::TINYINT:
-		case LogicalTypeId::SMALLINT:
-		case LogicalTypeId::INTEGER:
-		case LogicalTypeId::BIGINT:
-		case LogicalTypeId::HUGEINT: {
-			ScalarFunction function({type}, type, ScalarFunction::GetScalarUnaryFunction<TryAbsOperator>(type));
-			function.SetStatisticsCallback(PropagateAbsStats);
-			abs.AddFunction(function);
-			break;
-		}
-		case LogicalTypeId::FLOAT:
-		case LogicalTypeId::DOUBLE: {
-			ScalarFunction function({type}, type, ScalarFunction::GetScalarUnaryFunction<AbsOperator>(type));
-			function.SetStatisticsCallback(PropagateAbsStats);
-			abs.AddFunction(function);
-			break;
-		}
-		case LogicalTypeId::UTINYINT:
-		case LogicalTypeId::USMALLINT:
-		case LogicalTypeId::UINTEGER:
-		case LogicalTypeId::UBIGINT:
-			abs.AddFunction(ScalarFunction({type}, type, ScalarFunction::NopFunction));
-			break;
-		default:
-			abs.AddFunction(ScalarFunction({type}, type, ScalarFunction::GetScalarUnaryFunction<AbsOperator>(type)));
-			break;
-		}
-	}
-	abs.SetFallible();
-	NumericFunctionIdentity::PreserveStatistics(abs);
-	return abs;
 }
 
 //===--------------------------------------------------------------------===//
@@ -476,7 +291,6 @@ ScalarFunctionSet SignFun::GetFunctions() {
 		ScalarFunction function({type}, LogicalType::TINYINT,
 		                        ScalarFunction::GetScalarUnaryFunctionFixedReturn<int8_t, SignOperator>(type));
 		function.SetStatisticsCallback(PropagateSignStats);
-		NumericFunctionIdentity::PreserveStatistics(function);
 		sign.AddFunction(function);
 	}
 	return sign;
@@ -1352,7 +1166,6 @@ ScalarFunction PowOperatorFun::GetFunction() {
 	                        BindIEEEFloatingBinary<PowOperator, IEEEPowOperator>);
 	function.SetStatisticsCallback(PropagatePowStats);
 	function.SetFallible();
-	NumericFunctionIdentity::PreserveStatistics(function);
 	return function;
 }
 
@@ -1649,12 +1462,10 @@ ScalarFunctionSet IsNanFun::GetFunctions() {
 	ScalarFunction float_function({LogicalType::FLOAT}, LogicalType::BOOLEAN,
 	                              ScalarFunction::UnaryFunction<float, bool, IsNanOperator>);
 	float_function.SetStatisticsCallback(PropagateIsNanStats);
-	NumericFunctionIdentity::PreserveStatistics(float_function);
 	funcs.AddFunction(float_function);
 	ScalarFunction double_function({LogicalType::DOUBLE}, LogicalType::BOOLEAN,
 	                               ScalarFunction::UnaryFunction<double, bool, IsNanOperator>);
 	double_function.SetStatisticsCallback(PropagateIsNanStats);
-	NumericFunctionIdentity::PreserveStatistics(double_function);
 	funcs.AddFunction(double_function);
 	return funcs;
 }
