+++ src/function/aggregate/distributive/CMakeLists.txt
@@ -1,5 +1,11 @@
-add_library_unity(duckdb_aggr_distr OBJECT count.cpp decimal_avg.cpp
-                  first_last_any.cpp minmax.cpp)
+add_library_unity(
+  duckdb_aggr_distr
+  OBJECT
+  count.cpp
+  decimal_avg.cpp
+  first_last_any.cpp
+  minmax.cpp
+  sum.cpp)
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_aggr_distr>
     PARENT_SCOPE)
