+++ src/function/scalar/operator/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_library_unity(
   duckdb_func_ops_main
   OBJECT
+  abs.cpp
   add.cpp
   arithmetic.cpp
   cast.cpp
