+++ src/function/scalar/variant/variant_comparator.cpp
@@ -1,5 +1,4 @@
 #include "duckdb/function/scalar/variant_functions.hpp"
-#include "function_identity.hpp"
 #include "duckdb/common/radix.hpp"
 #include "duckdb/common/types/datetime.hpp"
 #include "duckdb/common/types/interval.hpp"
@@ -14,12 +13,6 @@
 #include "duckdb/storage/statistics/base_statistics.hpp"
 
 namespace duckdb {
-
-struct VariantComparatorFunctionIdentity {
-	static void PreserveStatistics(ScalarFunction &function) {
-		FunctionIdentityPreservation::PreserveStatistics(function);
-	}
-};
 
 namespace {
 
@@ -361,7 +354,6 @@ ScalarFunction VariantComparatorFun::GetFunction() {
 	auto variant_type = LogicalType::VARIANT();
 	ScalarFunction function("variant_comparator", {variant_type}, LogicalType::BLOB, VariantComparatorFunction);
 	function.SetStatisticsCallback(VariantComparatorStats);
-	VariantComparatorFunctionIdentity::PreserveStatistics(function);
 	return function;
 }
