+++ src/include/duckdb/function/built_in_functions.hpp
@@ -54,6 +54,8 @@ private:
 	void RegisterArrowFunctions();
 	void RegisterSnifferFunction();
 	void RegisterCopyFunctions();
+	void RegisterAbsoluteValueFunctions();
+	void RegisterSumFunctions();
 
 	void RegisterExtensionOverloads();
