+++ src/include/duckdb/main/extension_entries.hpp
@@ -51,7 +51,6 @@ static constexpr ExtensionFunctionEntry EXTENSION_FUNCTIONS[] = {
     {"<@", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
     {">>", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
     {">>=", "inet", CatalogType::SCALAR_FUNCTION_ENTRY},
-    {"@", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
     {"@>", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
     {"^", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
     {"^@", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
@@ -61,7 +60,6 @@ static constexpr ExtensionFunctionEntry EXTENSION_FUNCTIONS[] = {
     {"__aws__cloudformation__quack_on_ec2__status", "aws", CatalogType::TABLE_FUNCTION_ENTRY},
     {"__internal_json_copy_to_geojson", "json", CatalogType::SCALAR_FUNCTION_ENTRY},
     {"__internal_json_copy_to_json", "json", CatalogType::SCALAR_FUNCTION_ENTRY},
-    {"abs", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
     {"acos", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
     {"acosh", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
     {"add_numbers_together", "demo_capi", CatalogType::SCALAR_FUNCTION_ENTRY},
@@ -966,8 +964,6 @@ static constexpr ExtensionFunctionEntry EXTENSION_FUNCTIONS[] = {
     {"struct_keys", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
     {"struct_update", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
     {"struct_values", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
-    {"sum", "core_functions", CatalogType::AGGREGATE_FUNCTION_ENTRY},
-    {"sum_no_overflow", "core_functions", CatalogType::AGGREGATE_FUNCTION_ENTRY},
     {"sumkahan", "core_functions", CatalogType::AGGREGATE_FUNCTION_ENTRY},
     {"switch", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
     {"tan", "core_functions", CatalogType::SCALAR_FUNCTION_ENTRY},
